framework: cleanup for carthage.
