--- Framework/JWT_iOS_Framework/JWT_iOS_Framework.h
@@ -1,19 +1,0 @@
-//
-//  JWT_iOS_Framework.h
-//  JWT_iOS_Framework
-//
-//  Created by Lobanov Dmitry on 12.08.16.
-//  Copyright © 2016 Karma. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for JWT_iOS_Framework.
-FOUNDATION_EXPORT double JWT_iOS_FrameworkVersionNumber;
-
-//! Project version string for JWT_iOS_Framework.
-FOUNDATION_EXPORT const unsigned char JWT_iOS_FrameworkVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <JWT_iOS_Framework/PublicHeader.h>
-
-
