--- Framework/JWT_iOS_Static/JWT_iOS_Static.h
@@ -1,13 +1,0 @@
-//
-//  JWT_iOS_Static.h
-//  JWT_iOS_Static
-//
-//  Created by Lobanov Dmitry on 12.08.16.
-//  Copyright © 2016 Karma. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-
-@interface JWT_iOS_Static : NSObject
-
-@end
