--- Framework/JWT_iOS_Swift/JWT_iOS_Swift.h
@@ -1,19 +1,0 @@
-//
-//  JWT_iOS_Swift.h
-//  JWT_iOS_Swift
-//
-//  Created by Lobanov Dmitry on 12.08.16.
-//  Copyright © 2016 Karma. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for JWT_iOS_Swift.
-FOUNDATION_EXPORT double JWT_iOS_SwiftVersionNumber;
-
-//! Project version string for JWT_iOS_Swift.
-FOUNDATION_EXPORT const unsigned char JWT_iOS_SwiftVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <JWT_iOS_Swift/PublicHeader.h>
-
-
