--- Framework/JWT_macOS_Framework/JWT_macOS_Framework.h
@@ -1,19 +1,0 @@
-//
-//  JWT_macOS_Framework.h
-//  JWT_macOS_Framework
-//
-//  Created by Lobanov Dmitry on 12.08.16.
-//  Copyright © 2016 Karma. All rights reserved.
-//
-
-#import <Cocoa/Cocoa.h>
-
-//! Project version number for JWT_macOS_Framework.
-FOUNDATION_EXPORT double JWT_macOS_FrameworkVersionNumber;
-
-//! Project version string for JWT_macOS_Framework.
-FOUNDATION_EXPORT const unsigned char JWT_macOS_FrameworkVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <JWT_macOS_Framework/PublicHeader.h>
-
-
