--- Framework/JWT_macOS_Swift/JWT_macOS_Swift.h
@@ -1,19 +1,0 @@
-//
-//  JWT_macOS_Swift.h
-//  JWT_macOS_Swift
-//
-//  Created by Lobanov Dmitry on 12.08.16.
-//  Copyright © 2016 Karma. All rights reserved.
-//
-
-#import <Cocoa/Cocoa.h>
-
-//! Project version number for JWT_macOS_Swift.
-FOUNDATION_EXPORT double JWT_macOS_SwiftVersionNumber;
-
-//! Project version string for JWT_macOS_Swift.
-FOUNDATION_EXPORT const unsigned char JWT_macOS_SwiftVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <JWT_macOS_Swift/PublicHeader.h>
-
-
