+++ Framework/Supplement/JWT.modulemap
@@ -1,0 +1,5 @@
+framework module JWT {
+	umbrella header "JWT.h"
+	export *
+	module * { export * }
+}
